refactor(server): consolidate duplicate utils into server/utils/
Three categories of duplicated utility functions moved to shared
modules under server/utils/:

1. makeId() — was copy-pasted in schedulerHandlers, todosHandlers,
   todosItemsHandlers (3 copies). Consolidated to server/utils/id.ts
   with a required `prefix` param so IDs remain distinguishable
   across domains (todo_*, sched_*).

2. isValidSlug() — identical 6-line function in sources/paths.ts
   and skills/paths.ts. Moved to server/utils/slug.ts alongside
   the existing slugify/hashSlug helpers. All consumers + tests
   updated to import from the new location.

3. formatSpawnFailure() + extractClaudeErrorMessage() — near-
   identical spawn-error formatting in chat-index/summarizer,
   sources/pipeline/summarize, and sources/classifier (3 copies,
   ~45 lines each). Consolidated to server/utils/spawn.ts. Each
   call site now passes its log prefix ("[chat-index]", etc.)
   as a parameter instead of baking it in.

Also adds the "Utility functions: group by concern" guideline to
CLAUDE.md so future code follows the same pattern.

19 files changed; net reduction ~120 lines of duplicate code.
